Restructure tests
For #77

Put different directory for tests that depend on sklearn, keras, pytorch

This is to allow testing with some of the packages uninstalled.

Still not working as some keras and pytorch depend on an sklearn
dataset.
