+++ tests/pytorch/test_exceptions.py
