+++ tests/sklearn/test_exceptions.py
