+++ tests/sklearn/test_register_constr.py
